[FEAT] Article 엔터티 이미지 관리 및 embeddable ID 개선 src/main/java/com/teambind/articleserver/entity/embeddable_id/ArticleImagesId.java: sequenceNum 필드 타입 int → Long으로 변경 src/main/java/com/teambind/articleserver/entity/Article.java: 이미지 추가/삭제 메서드(`addImage`, `removeImages`) 구현 및 `addKeyword` 로직 수정 src/main/java/com/teambind/articleserver/entity/ArticleImage.java: 이미지 엔터티 생성자 추가
참고: DB 마이그레이션 필요

+++ src/main/java/com/teambind/articleserver/entity/embeddable_id/ArticleImagesId.java
@@ -16,9 +16,9 @@ public class ArticleImagesId implements Serializable {
 	@Column(name = "article_id", nullable = false)
 	private String articleId;
 	@Column(name = "sequence_no", nullable = false)
-	private int sequenceNum;
+	private Long sequenceNum;
 	
-	public ArticleImagesId(String articleId, int sequenceNum) {
+	public ArticleImagesId(String articleId, Long sequenceNum) {
 		this.articleId = articleId;
 		this.sequenceNum = sequenceNum;
 	}
